finished get cell_ref + need to think if i return the correct type + need to continue with methods

+++ src/page/mod.rs
@@ -22,7 +22,7 @@ pub(crate) type RawPage = [u8; 4096];
 pub(crate) struct PageID(pub u64);
 
 #[derive(Eq, Hash, PartialEq, Debug)]
-pub(crate) struct SlotID(pub u64);
+pub(crate) struct SlotID(pub u16);
 
 #[derive(Debug)]
 #[derive(PartialEq)]
